pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bas-orchestrator"
version = "0.1.0"
description = "Campaign orchestrator for safe BAS simulations with evidence packs."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Sarvesh Kapre" }]
dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.10",
  "types-PyYAML>=6.0",
  "pip-audit>=2.7",
  "build>=1.2",
  "bandit>=1.7",
]

[project.scripts]
bas = "bas_orchestrator.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true

[tool.pytest.ini_options]
addopts = "-q"
